requirements.txt
streamlit==1.32.0
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
python-dotenv==1.0.0
